requirements.txt
alembic==1.13.1
aiofiles==23.2.1
aiomysql==0.2.0
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
bcrypt==4.0.1
blinker==1.7.0
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
cssselect2==0.7.0
dnspython==2.6.1
email_validator==2.1.1
et-xmlfile==1.1.0
fastexcel==0.10.2
fastapi==0.112.4
Flask==3.0.3
fonttools==4.51.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
loguru==0.7.2
Mako==1.3.2
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.26.4
openpyxl==3.1.2
pandas==2.1.4
passlib==1.7.4
Pillow==10.2.0
polars==0.20.16
pyarrow==15.0.2
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.4
pydantic-core==2.16.3
pydantic-settings==2.1.0
pydyf==0.8.0
PyJWT==2.8.0
PyMySQL==1.1.0
pyphen==0.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rich==13.7.1
# rignore removed - package not found
rsa==4.9
sentry-sdk==1.45.0
setuptools==69.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.38.2
tinycss2==1.2.1
tinyhtml5==2.0.0
# typing-inspection removed - deprecated package
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
websockets==12.0
Werkzeug==3.0.1
wheel==0.43.0
win32-setctime==1.1.0
XlsxWriter==3.2.0
zopfli==0.2.3

# Testing dependencies
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0
pytest-xdist==3.6.1
pytest-mock==3.14.0
httpx==0.27.2
aiosqlite==0.20.0

# Code quality
flake8==7.1.1
black==24.8.0
isort==5.13.2
mypy==1.11.2

# Type stubs for mypy
types-aiofiles==23.2.0
types-python-dateutil==2.9.0.20240316
types-pytz==2024.1.0.20240417
types-requests==2.32.0.20240622
types-PyYAML==6.0.12.20240311

# MikroTik API
routeros-api==0.21.0
netmiko==4.2.0

# Security
bandit==1.7.6
safety==3.2.0
